graph: Move entity caching logic to separate file
